+++ GANDLF/cli/__init__.py
@@ -1,3 +1,13 @@
 from .patch_extraction import patch_extraction
 from .main_run import main_run
 from .preprocess_and_save import preprocess_and_save
+from .config_generator import config_generator
+
+from datetime import date
+
+copyrightMessage = (
+    "Contact: gandlf@mlcommons.org\n\n"
+    + "This program is NOT FDA/CE approved and NOT intended for clinical use.\nCopyright (c) "
+    + str(date.today().year)
+    + " MLCommons. All rights reserved.\n\nCitation: https://arxiv.org/abs/2103.01006"
+)