+++ GANDLF/models/light_unet.py
@@ -29,11 +29,10 @@ class light_unet(ModelBase):
 
         self.network_kwargs = {"res": False}
 
-        if not (checkPatchDivisibility(parameters["patch_size"])):
-            sys.exit(
-                "The patch size is not divisible by 16, which is required for",
-                parameters["model"]["architecture"],
-            )
+        assert checkPatchDivisibility(parameters["patch_size"]) == True, (
+            "The patch size is not divisible by 16, which is required for "
+            + parameters["model"]["architecture"]
+        )
 
         self.ins = in_conv(
             input_channels=self.n_channels,