+++ HISTORY.md
@@ -2,6 +2,7 @@
 - ImageNet pre-trained models for UNet with variable encoders is now available
 - ACS/Soft conversion is available for ImageNet-pretrained UNet
 - Updated links, copyright and email to MLCommons
+- Allowing provision for user to generate multiple configurations for experimentation
 
 ## 0.0.15
 - Updated `setup.py` for `python>=3.8`